### 2023-12-09 #6135 @editorialbot
[PRE REVIEW]: CCS-Lib: A Python package to elicit latent knowledge from LLMs
**Submitting author:** @lauritowal (<a href="http://orcid.org/0000-0002-6431-7053">Walter Laurito</a>)
**Repository:** https://github.com/EleutherAI/ccs
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.1.1
