### 2023-12-14 #6135 @lauritowal
Hi @arfon - thanks. Is it possible to change the repository URL to: https://github.com/EleutherAI/ccs and the title to CCS-Lib: A Python package to elicit latent knowledge from LLMs
Everything else stays the same. Thanks!
