### 2024-01-24 #6135 @mooniean
Hi @lauritowal!

The project repository changed, as well as the name of the package - could you update the paper with these new nomenclatures please? I can only find versions with the previous name!
