### 2024-01-26 #6135 @mooniean
@lauritowal  thank you for changing! The name of the paper still doesn't match the one requested here (CCS-Lib: A Python package to elicit later knowledge from LLMs) - fixing that should be ok!

The references are not formatted properly (see [here](https://joss.readthedocs.io/en/latest/submitting.html#citations)), if you could re-arrange that then all is ok!

Is there any reviewers you'd like to suggest?
