+++ joss/paper.bib
@@ -1,16 +1,14 @@
-@article{burns,
+@article{burns2022,
   title={Discovering latent knowledge in language models without supervision},
   author={Burns, Collin and Ye, Haotian and Klein, Dan and Steinhardt, Jacob},
   journal={arXiv preprint arXiv:2212.03827},
   year={2022}
 }
 
-@misc{elk,
+@misc{christiano2021,
 	author = {Christiano, Paul and Cotra, Ajeya and Xu, Mark},
 	title = {Eliciting Latent Knowledge (ELK)},
 	howpublished = {\url{https://docs.google.com/document/d/1WwsnJQstPq91_Yh-Ch2XRL8H_EpsnjrC1dwZXR37PC8/}},
 	year = {December 2021},
 	note = {[Accessed 11-08-2023]},
 }
-
-
