### 2024-02-22 #6135 @mooniean
Hi 👋 @patrickraoulphilipp @guanqun-yang @jhagerer - would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> CCS-Lib: A Python package to elicit latent knowledge from LLMs #6135

Thanks in advance!
