### 2024-03-07 #6135 @mooniean
Hi [@harishsiravuri](https://github.com/harishsiravuri) [@praneethd7](https://github.com/praneethd7) would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> CCS-Lib: A Python package to elicit latent knowledge from LLMs #6135

Thanks in advance!
