### 2024-03-08 #6135 @mooniean
Oops I think I didn't tag them and instead put a link. Re-retrial:

Hi @harishsiravuri @praneethd7 would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> CCS-Lib: A Python package to elicit latent knowledge from LLMs https://github.com/openjournals/joss-reviews/issues/6135

Thanks in advance!
