### 2024-03-21 #6135 @mooniean
Hi @isdanni @Saran-nns ! Would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> CCS-Lib: A Python package to elicit latent knowledge from LLMs #6135

Thanks in advance!
