### 2024-03-21 #6511 @mooniean
Hi  @lauritowal @isdanni @praneethd7 - this is the review thread for the submission. All of our communications will happen here from now on. The title was updated on the pre-review but I'm not sure why it wasn't updated here - I'll try to sort it!

Meanwhile, please check the post at the top of the issue for instructions on how to generate your own review checklist. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
