### 2024-03-22 EleutherAI/ccs#8 @isdanni
[JOSS Review] elk: A Python package to elicit latent knowledge from LLMs
tracking review progress: https://github.com/openjournals/joss-reviews/issues/6511

Note: this is a review thread, more checklist items will be added.

## Documentation

- [ ] In the https://github.com/EleutherAI/ccs/tree/main?tab=readme-ov-file#quick-start section, it is usually recommended to install all the required packages in a specific virtual environment instead of system-wide.
- [ ]  Please merge https://github.com/EleutherAI/ccs/tree/main?tab=readme-ov-file#development and https://github.com/EleutherAI/ccs/tree/main?tab=readme-ov-file#contributing-to-this-repository to **Contribution Guidelines**. It should at least include answers to 3 questions in the JOSS review principles: *Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support*. Here are some useful templates(There's no need to include everything but this section should at least include the basic answers. For example, how to submit an issue, how to create a pull request and how to set up the development environment, and is there any difference from the production or main branch): https://github.com/auth0/open-source-template/blob/master/GENERAL-CONTRIBUTING.md, https://gist.github.com/briandk/3d2e8b3ec8daf5a27a62, https://github.com/pytorch/pytorch/blob/main/CONTRIBUTING.md, https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors.
- [ ] Are there any examples for using the codebase? With real-world analysis questions preferably.

## Minor issues

- [ ] In [LICENSE.md](https://github.com/EleutherAI/ccs/blob/main/LICENSE.md), please update the year to 2024.
- [ ] The bold markdown syntax did not work when an URL is wrapped around the text.
![Screenshot from 2024-03-22 00-36-34](https://github.com/EleutherAI/ccs/assets/38550500/c4ae2eec-66eb-4687-9aee-08919c7950e7)
