### 2024-03-22 EleutherAI/ccs#8 @isdanni
## Tests

Please ignore the previous comments. Using `pip install -e .` solved the package issue.

- [x]  When running `pytest` on a single-core machine, there are 5 failed tests related to not having more than 1 GPU. If the elk package can only run in a multi-core environment, please specify it in the documentation.

![Screenshot from 2024-03-22 00-55-50](https://github.com/EleutherAI/ccs/assets/38550500/66169ac3-f6b9-4695-abb8-8fcef1540f6d)
