### 2024-03-25 #6511 @praneethd7
Hello Walter @lauritowal! I just started the review process.  I wanted to get some clarity on the following checkbox item:

>  Contribution and authorship: Has the submitting author (@lauritowal) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

I am curious that there are more contributors on the GitHub repository than the number of authors listed in the submission. Could you confirm if you checked with all the contributors regarding the authorship? Perhaps, some of their contributions are minor but I wanted to make sure.
