### 2024-04-19 EleutherAI/ccs#8 @isdanni
## Paper

- [x] There's a typo in the header column: "machine leaarning"
- [x] Please elaborate on the "enhanced version of the CSS method" and how it compares to existing implementations/repositories(if any).
- [x] More details of the VINC method from the original paper. For example, why it's important, who would use it and how it's different from the other methods.
- [x] Is  Contrastive Representation Clustering (CRC) listed in References?
- [x] "Statement of the Need" could use a more in-depth introduction
