### 2024-09-02 EleutherAI/ccs#8 @lauritowal
@isdanni I've done the changes and merged the latest main into the joss-paper branch. The pull-request will merged soon: https://github.com/EleutherAI/ccs/pull/9
