### 2024-09-03 EleutherAI/ccs#8 @isdanni
> @isdanni I've done the changes and merged the latest main into the joss-paper branch. The pull-request will merged soon: #9

Thanks for the update! I checked off most of the task items. Just one last thing(very minor issue):

In the Introduction section of README.md, the bold syntax does not work with URL; Unless this is intended I'd suggest removing it for better readability.
![Screenshot from 2024-09-02 21-55-45](https://github.com/user-attachments/assets/3e92d2fa-6687-4341-90fd-a681827572b6)
