### 2024-09-03 #6511 @isdanni
Thanks @lauritowal for the update! I finished the initial round of the review at https://github.com/EleutherAI/ccs/issues/8 and checked off the list items
