### 2024-09-04 EleutherAI/ccs#11 @praneethd7
[JOSS Review] Import Errors
Hey @lauritowal

Hope all is good. Firstly, my apologies for not being timely on the review. Also, I had few overlapping suggestions with the other reviewer so I was waiting on the changes. I was trying to test the functionality and I couldn't get it running right off the bat.
![image](https://github.com/user-attachments/assets/87da4de8-019d-4612-a8cb-30035e3e4cff)

Could you help me with this? Also, it would be nice to have an example notebook to get started. You could also have it as Colab notebook.

for reference: https://github.com/openjournals/joss-reviews/issues/6511
