### 2024-09-04 EleutherAI/ccs#11 @lauritowal
Hi @praneethd7 - thanks for taking the time to review!
I've added a small example notebook into the joss-paper branch: https://github.com/EleutherAI/ccs/tree/joss-paper
The example is linked in the Readme in that branch.

Since, I've added other changes based on a previous review to that branch, it would be best if you look at that for your review instead of main. We will later merge the branch into main :)
