### 2025-03-17 #6511 @praneethd7
@lauritowal Thank you. I do not see any changes. The last commit on https://github.com/EleutherAI/ccs/tree/joss-paper is from 6 months ago. Can you check if you have pushed the changes?
