### 2025-03-25 #6511 @mooniean
Line 39: the ccs in "ccs library" doesn't have the code formatting that it has in previous uses
Line 99: "it easy to compare the performance of for a given model" > either of or for, not both :)
Line 105: citation to be inside parenthesis

There's inconsistency throughout with HuggingFace/Huggingface, please change them all to be consistent to one of the formattings!

After these, I'm happy to proceed once all the archiving is done!
