### 2025-04-14 #6511 @lauritowal
Thanks!

@mooniean could you recreate the PDF please from the new repository? I would like to make sure that all works and that my changes are there :)
