### 2025-10-17 #6511 @crvernon
@lauritowal - Please address the following before I accept this for publication:

- [ ] LINES 41-42:  You already define the acronym and cite this work earlier in the paper:  `ELK (Christiano et al., December 2021))`.  There is not need to do it here again.
- [ ] LINES 46-47:  You state, "...method described in the paper “Discovering Latent Knowledge in
47 Language Models Without Supervision” by Burns et al. (2022)."  There is no need to state the title of the paper here.  Simply state, "method described in Burns et al. (2022)."

Let me know when these changes have been addressed and I'll accept.  Thank you.
